backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
zhipuai>=2.0.0
arxiv>=2.0.0
scholarly>=1.7.0
duckduckgo-search>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
